requirements.txt
anthropic
google-api-python-client
google-auth
matplotlib
numpy
pydantic
python-docx
python-dotenv
